Major update (Read Comments) Login Page
Sorry for the update dump.
Modulised navigation menu to be a php file to be loaded by all pages, changed directory name "html -> content" because all files are php files now.
Modified login page to use bootstrap and the same style as the rest of the
web page... it's currently incomplete as the function to login asks for
email and pw, and we may prefer using a username instead (write a
get_email function next).
Set the navigation menu to show "logout" if we are logged in, and vice
versa (not yet complete).